balance: P0-1 需求系数降低 + 升级订单修复
需求系数: mechanized 2→1.5, steam 3→2.5, electrified 4→3.5
- 升级现在产生surplus goods可卖海外
- 升级mechanized回本周期约1.7回合

升级订单修复:
- _normalize_upgrade_orders: 支持routeId和targetMode两种格式
- resolver: 从factoryPlan读取upgradeOrders（原来从phase1Production读导致静默跳过）
- 303 passed, 0 regression

+++ backend/app/modules/game_state/workspaces.py
@@ -332,7 +332,7 @@ def _build_phase1_production_modes(player: PlayerState) -> list[dict[str, Any]]:
                 "label": PHASE1_MODE_LABELS.get(mode, mode),
                 "inputRatio": 1,
                 "outputRatio": int(PRODUCTION_MODE_OUTPUT_RATIOS[mode]),
-                "demandCoefficient": int(PRODUCTION_MODE_DEMAND_COEFFICIENTS[mode]),
+                "demandCoefficient": float(PRODUCTION_MODE_DEMAND_COEFFICIENTS[mode]),
                 "buildCost": int(new_factory_costs.get(mode, 0)),
                 "upgradeCost": int(upgrade_costs.get(mode, 0)),
                 "currentCapacity": int(capacity_by_mode.get(mode, 0)),
